Making the error boundary text translatable
