+++ packages/dynamic-flows/i18n/en.json
@@ -1,4 +1,5 @@
 {
+  "dynamicFlows.ErrorBoundary.errorAlert": "Oops. Something went wrong...",
   "dynamicFlows.Help.ariaLabel": "Click here for more info.",
   "dynamicFlows.PersistAsyncSchema.genericError": "Something went wrong, please try again later!"
 }
