+++ packages/dynamic-flows/i18n/index.js
@@ -1,4 +1,31 @@
 import en from './en.json';
+import de from './de.json';
+import es from './es.json';
+import fr from './fr.json';
+import hu from './hu.json';
+import id from './id.json';
+import it from './it.json';
 import ja from './ja.json';
+import pl from './pl.json';
+import pt from './pt.json';
+import ro from './ro.json';
+import ru from './ru.json';
+import tr from './tr.json';
+import zh from './zh.json';
 
-export default { en, ja };
+export default {
+  en,
+  de,
+  es,
+  fr,
+  hu,
+  id,
+  it,
+  ja,
+  pl,
+  pt,
+  ro,
+  ru,
+  tr,
+  zh,
+};
