+++ packages/dynamic-flows/src/flow/errorBoundary/ErrorBoundary.messages.js
@@ -1,0 +1,9 @@
+import { defineMessages } from 'react-intl';
+
+export default defineMessages({
+  errorAlert: {
+    id: 'dynamicFlows.ErrorBoundary.errorAlert',
+    defaultMessage: 'Oops. Something went wrong...',
+    description: 'Description of Dynamic flows error alert',
+  },
+});
